+++ example/guard.ts
@@ -1,22 +1,43 @@
-import { z } from 'zod'
-import KingWorld from '../src'
+import KingWorld, { KingWorldInstance, TypedSchema } from '../src'
+import { z, ZodObject } from 'zod'
+
+const a = z.object({
+	name: z.string()
+})
 
 new KingWorld()
-	.state('awd', 'a')
-	.get('/', () => 'Hi')
+	.state('name', 'salt')
+	.get('/', ({ store: { name } }) => `Hi ${name}`, {
+		schema: {
+			query: z.object({
+				name: z.string()
+			})
+		}
+	})
+	// If query 'name' is not preset, skip the whole handler
 	.guard(
 		{
-			beforeHandle: ({ query }) => {
-				if (!query.name) return 'Query name is required'
+			schema: {
+				query: z.object({
+					name: z.string()
+				})
 			}
 		},
 		(app) =>
-			app.get('/profile', ({ query: { name } }) => `Hi ${name}`, {
-				schema: {
-					query: z.object({
-						name: z.string()
-					})
-				}
-			})
+			app
+				// Query type is inherited from guard
+				.get('/profile', ({ query: { name } }) => `Hi ${name}`)
+				// Store is inherited
+				.post('/name', ({ store: { name } }) => name, {
+					schema: {
+						body: z.object({
+							id: z.number().min(5),
+							username: z.string(),
+							profile: z.object({
+								name: z.string()
+							})
+						})
+					}
+				})
 	)
 	.listen(3000)